Rename noisy interface
